refactor(integrations): share agntcy-a2a interop behaviors
