+++ integrations/agntcy-a2a/README.md
@@ -8,7 +8,7 @@ All 12 Rust/Go client-server legs in the current core lifecycle matrix are green
 
 The Go and Rust fixtures now expose push-config CRUD. CSIT validates that path from both clients against both server targets across JSON-RPC, HTTP+JSON, and gRPC.
 
-The Rust/.NET slice now exists alongside the Rust/Go suite. It reuses the existing Rust fixture and Rust probe, adds CSIT-owned .NET fixture and probe binaries, and runs exactly 8 specs: 4 JSON-RPC legs (`dotnet-dotnet`, `dotnet-rust`, `rust-dotnet`, `rust-rust-dotnet`) plus the same 4 legs over HTTP+JSON. Each of those specs now exercises the original lifecycle slice plus the richer scenario-parity slice: message-only responses, failed tasks, multi-turn input-required continuations, long-running completions, artifact-rich streaming updates, structured text/data/URL artifacts, extended-agent-card discovery, and preservation of mixed text plus structured-data content through task history. Push-config is covered in this slice as well, but only the Rust-server legs are expected to support push-config CRUD; the .NET-server legs are expected to return the current unsupported error. This slice does not cover gRPC. The default `task test` and `task integrations:a2a:test` entrypoints now run both the established Rust/Go matrix and this Rust/.NET slice.
+The Rust/.NET slice now exists alongside the Rust/Go suite. It reuses the existing Rust fixture and Rust probe, adds CSIT-owned .NET fixture and probe binaries, and covers 8 legs: 4 JSON-RPC legs (`dotnet-dotnet`, `dotnet-rust`, `rust-dotnet`, `rust-rust-dotnet`) plus the same 4 legs over HTTP+JSON. Each leg is split into shared behavior specs for unary and streaming requests, task lifecycle APIs, push-config semantics, and scenario parity. Push-config is covered in this slice as well, but only the Rust-server legs are expected to support push-config CRUD; the .NET-server legs are expected to return the current unsupported error. This slice does not cover gRPC. The default `task test` and `task integrations:a2a:test` entrypoints now run both the established Rust/Go matrix and this Rust/.NET slice.
 
 Across the matrix, the scenarios validate the same core interoperability behavior:
 
@@ -27,7 +27,7 @@ Across the matrix, the scenarios validate the same core interoperability behavio
 
 The fixtures are intentionally small and deterministic so the suite can run the same way locally and in CI without depending on sibling SDK checkouts.
 
-Each scenario is tagged with a dedicated Ginkgo label and exposed through a matching Task target so the full matrix and each individual leg can be run independently.
+Each leg keeps its existing suite, transport, and pair labels for Task targets, and the shared behavior slices add dedicated Ginkgo labels (`behavior-core`, `behavior-unary-streaming`, `behavior-lifecycle`, `behavior-push-config`, and `behavior-parity`) so full legs and cross-leg behavior groups can be filtered independently.
 
 The gRPC legs follow the same agent-card discovery path as the other transports: each fixture serves `/.well-known/agent-card.json` over HTTP and advertises a separate gRPC transport endpoint from that card.
 
@@ -120,12 +120,28 @@ task test:rust-dotnet:jsonrpc
 task test:rust-dotnet:jsonrpc:dotnet-dotnet
 task test:rust-dotnet:jsonrpc:dotnet-rust
 task test:rust-dotnet:jsonrpc:rust-dotnet
-task test:rust-dotnet:jsonrpc:rust-rust
+task test:rust-dotnet:jsonrpc:rust-rust-dotnet
 task test:rust-dotnet:rest
 task test:rust-dotnet:rest:dotnet-dotnet
 task test:rust-dotnet:rest:dotnet-rust
 task test:rust-dotnet:rest:rust-dotnet
-task test:rust-dotnet:rest:rust-rust
+task test:rust-dotnet:rest:rust-rust-dotnet
+```
+
+Behavior-focused runs are exposed as first-class Task targets as well:
+
+```sh
+task test:behavior:core
+task test:behavior:unary-streaming
+task test:behavior:lifecycle
+task test:behavior:push-config
+task test:behavior:parity
+task test:rust-go:behavior:unary-streaming
+task test:rust-go:behavior:lifecycle
+task test:rust-go:behavior:push-config
+task test:rust-dotnet:behavior:unary-streaming
+task test:rust-dotnet:behavior:lifecycle
+task test:rust-dotnet:behavior:push-config
 ```
 
 From the repository root:
@@ -160,12 +176,25 @@ task integrations:a2a:test:rust-dotnet:jsonrpc
 task integrations:a2a:test:rust-dotnet:jsonrpc:dotnet-dotnet
 task integrations:a2a:test:rust-dotnet:jsonrpc:dotnet-rust
 task integrations:a2a:test:rust-dotnet:jsonrpc:rust-dotnet
-task integrations:a2a:test:rust-dotnet:jsonrpc:rust-rust
+task integrations:a2a:test:rust-dotnet:jsonrpc:rust-rust-dotnet
 task integrations:a2a:test:rust-dotnet:rest
 task integrations:a2a:test:rust-dotnet:rest:dotnet-dotnet
 task integrations:a2a:test:rust-dotnet:rest:dotnet-rust
 task integrations:a2a:test:rust-dotnet:rest:rust-dotnet
-task integrations:a2a:test:rust-dotnet:rest:rust-rust
+task integrations:a2a:test:rust-dotnet:rest:rust-rust-dotnet
+```
+
+The same behavior-focused targets are available from the repository root:
+
+```sh
+task integrations:a2a:test:behavior:core
+task integrations:a2a:test:behavior:unary-streaming
+task integrations:a2a:test:behavior:lifecycle
+task integrations:a2a:test:behavior:push-config
+task integrations:a2a:test:behavior:parity
+task integrations:a2a:test:rust-go:behavior:unary-streaming
+task integrations:a2a:test:rust-go:behavior:lifecycle
+task integrations:a2a:test:rust-dotnet:behavior:push-config
 ```
 
 Each run writes Ginkgo JSON and JUnit reports under `integrations/agntcy-a2a/reports/`. The combined Rust/Go suite emits `report-agntcy-a2a.{json,xml}`, the combined Rust/.NET suite emits `report-agntcy-a2a-rust-dotnet.{json,xml}`, the transport-scoped tasks emit `report-agntcy-a2a-jsonrpc.{json,xml}`, `report-agntcy-a2a-rest.{json,xml}`, `report-agntcy-a2a-grpc.{json,xml}`, `report-agntcy-a2a-rust-dotnet-jsonrpc.{json,xml}`, and `report-agntcy-a2a-rust-dotnet-rest.{json,xml}`, and the per-case tasks emit scenario-specific report names via `-ginkgo.label-filter`.
